@@ -28,9 +28,7 @@ module S3Browser
     before(:each) { allow(Bucket).to receive(:new).and_return(BucketDummy.new(bucket_name)) }
 
     describe "GET index" do
-      before  do
-        get :index, bucket_name: bucket_name
-      end
+      before  { get :index, params: { bucket_name: bucket_name }}
 
       it "renders the index template" do
         expect(response).to render_template("index")
@@ -39,14 +37,14 @@ module S3Browser
 
     describe "GET upload" do
       context 'when one or more files are selected' do
-        before { get :upload, bucket_name: bucket_name, files: ['array', 'of', 'files'] }
+        before { get :upload, params: { bucket_name: bucket_name, files: ['array', 'of', 'files'] }}
         it "redirects to the index template" do
           expect(response).to redirect_to(:action => "index", :bucket_name => 'foo-bucket')
         end
       end
 
       context 'when no file is selected' do
-        before { get :upload, bucket_name: bucket_name, files: nil }
+        before { get :upload, params: { bucket_name: bucket_name, files: nil }}
         it "rerenders the form with an error message" do
           expect(response).to render_template(:index)
           expect(assigns(:error_message)).to eq('Must select at least one file to upload.')
@@ -56,14 +54,14 @@ module S3Browser
 
     describe "GET delete" do
       context 'when a file is selected for deletion' do 
-        before { get :delete, bucket_name: bucket_name, filename_to_delete: 'foo_file' }
+        before { get :delete, params: { bucket_name: bucket_name, filename_to_delete: 'foo_file' }}
         it "redirects to the index template" do
           expect(response).to redirect_to(:action => "index", :bucket_name => 'foo-bucket')
         end
       end
 
       context 'when no file is selected' do
-        before { get :delete, bucket_name: bucket_name, filename_to_delete: nil }
+        before { get :delete, params: { bucket_name: bucket_name, filename_to_delete: nil }}
         it "rerenders the form with an error message" do
           expect(response).to render_template(:index)
           expect(assigns(:error_message)).to eq('Must select a file to delete.')
@@ -74,7 +72,7 @@ module S3Browser
     describe '#bucket' do 
       before  do
         allow(Bucket).to receive(:new).and_raise(StandardError)
-        get :index, bucket_name: bucket_name
+        get :index, params: { bucket_name: bucket_name }
       end
       it "renders an error message when there is trouble connecting to the bucket" do
         expect(response).to render_template(:error)
